+++ .gitignore
@@ -1,4 +1,4 @@
-/postviewd
+/appdetaild
 /.bin
 mocks
 .vscode/
