--- api/postview/postview.proto
@@ -1,37 +1,0 @@
-syntax = "proto3";
-
-// protoc -I. --go_out=plugins=grpc:$GOPATH/src ./api/postview/postview.proto
-
-package postview;
-option go_package = "github.com/cafebazaar/go-boilerplate/pkg/postview";
-
-import "google/protobuf/timestamp.proto";
-
-message GetPostRequest {
-    string token = 1;
-}
-
-message GetPostResponse {
-    Post post = 1;
-}
-
-message Post {
-    string title = 1;
-    string description = 2;
-    string token = 3;
-    string thumbnail = 4;
-    Contact contact = 5;
-    google.protobuf.Timestamp published_at = 6;
-    repeated string category_slugs = 7;
-}
-
-message Contact {
-    string phone = 1;
-    string email = 2;
-    bool chat = 4;
-}
-
-service PostView {
-    rpc GetPost(GetPostRequest) returns (GetPostResponse);
-}
-
