+++ cmd/appdetaild/logging.go
