+++ cmd/appdetaild/main.go
