+++ go.mod
@@ -1,25 +1,25 @@
-module github.com/cafebazaar/go-boilerplate
+module git.cafebazaar.ir/bardia/lazyapi
 
 require (
-	github.com/allegro/bigcache v1.2.1
-	github.com/axw/gocov v1.0.0 // indirect
 	github.com/certifi/gocertifi v0.0.0-20190506164543-d2eda7129713 // indirect
 	github.com/evalphobia/logrus_sentry v0.8.2
 	github.com/getsentry/raven-go v0.2.0
-	github.com/go-logfmt/logfmt v0.4.0 // indirect
-	github.com/go-redis/redis v6.15.2+incompatible
 	github.com/golang/protobuf v1.3.1
 	github.com/google/wire v0.3.0
 	github.com/grpc-ecosystem/go-grpc-middleware v1.0.0
 	github.com/grpc-ecosystem/go-grpc-prometheus v1.2.0
 	github.com/jinzhu/gorm v1.9.8
+	github.com/parnurzeal/gorequest v0.2.16
 	github.com/pkg/errors v0.8.0
-	github.com/prometheus/client_golang v0.9.4
+	github.com/prometheus/client_golang v0.9.4 // indirect
 	github.com/sirupsen/logrus v1.4.2
 	github.com/spf13/cobra v0.0.5
 	github.com/spf13/viper v1.4.0
 	github.com/stretchr/testify v1.3.0
-	golang.org/x/xerrors v0.0.0-20190513163551-3ee3066db522
+	golang.org/x/tools v0.0.0-20190422233926-fe54fb35175b
+	golang.org/x/xerrors v0.0.0-20191204190536-9bdfabe68543
 	google.golang.org/grpc v1.21.1
-	gopkg.in/gormigrate.v1 v1.5.0
+	moul.io/http2curl v1.0.0 // indirect
 )
+
+go 1.13
