--- internal/app/provider/errors.go
@@ -1,7 +1,0 @@
-package provider
-
-import "errors"
-
-var (
-	ErrNotFound = errors.New("post not found")
-)
