+++ internal/pkg/grpcserver/server.go
@@ -2,12 +2,12 @@ package grpcserver
 
 import (
 	"fmt"
-	"github.com/cafebazaar/go-boilerplate/pkg/errors"
-	"github.com/cafebazaar/go-boilerplate/pkg/postview"
+	"git.cafebazaar.ir/bardia/lazyapi/pkg/appdetail"
 	grpcmiddleware "github.com/grpc-ecosystem/go-grpc-middleware"
 	grpclogrus "github.com/grpc-ecosystem/go-grpc-middleware/logging/logrus"
 	grpcrecovery "github.com/grpc-ecosystem/go-grpc-middleware/recovery"
 	grpcprometheus "github.com/grpc-ecosystem/go-grpc-prometheus"
+	"github.com/pkg/errors"
 	"github.com/sirupsen/logrus"
 	"google.golang.org/grpc"
 	"net"
@@ -18,25 +18,24 @@ type Server struct {
 	server   *grpc.Server
 }
 
-func New(postViewServer postview.PostViewServer, logger *logrus.Logger, listenPort int) (*Server, error) {
+func New(appDetailServer appdetail.AppDetailServer, logger *logrus.Logger, listenPort int) (*Server, error) {
 	listener, err := net.Listen("tcp", fmt.Sprintf(":%d", listenPort))
 	if err != nil {
 		return nil, errors.Wrap(err, "failed to listen")
 	}
 
 	logEntry := logger.WithFields(map[string]interface{}{
-		"app": "postviewd",
+		"app": "appdetaild",
 	})
 
 	interceptors := []grpc.UnaryServerInterceptor{
 		grpclogrus.UnaryServerInterceptor(logEntry),
-		errors.UnaryServerInterceptor,
 		grpcprometheus.UnaryServerInterceptor,
 		grpcrecovery.UnaryServerInterceptor(),
 	}
 
 	grpcServer := grpc.NewServer(grpc.UnaryInterceptor(grpcmiddleware.ChainUnaryServer(interceptors...)))
-	postview.RegisterPostViewServer(grpcServer, postViewServer)
+	appdetail.RegisterAppDetailServer(grpcServer, appDetailServer)
 
 	return &Server{
 		listener: listener,
