--- internal/pkg/metrics/observer.go
@@ -1,6 +1,0 @@
-package metrics
-
-type Observer interface {
-	Observe(float64)
-	With(labels map[string]string) Observer
-}
