--- pkg/cache/layer.go
@@ -1,10 +1,0 @@
-package cache
-
-import "context"
-
-type Layer interface {
-	Get(ctx context.Context, key string) ([]byte, error)
-	Set(ctx context.Context, key string, value []byte) error
-	Delete(ctx context.Context, key string) error
-	Clear(ctx context.Context) error
-}
