--- pkg/sql/sql.go
@@ -1,12 +1,0 @@
-package sql
-
-type Config interface {
-	ConnectionString() string
-	Dialect() string
-	GetMaxIDLEConnection() int
-	GetMaxOpenConnection() int
-}
-
-type Migrater interface {
-	Migrate() error
-}
